+++ BraintreeCore/src/main/java/com/braintreepayments/api/AnalyticsClient.kt
@@ -169,11 +169,11 @@ internal class AnalyticsClient @VisibleForTesting constructor(
         try {
             val analyticsRequest = serializeEvents(authorization, events, metadata)
             httpClient.post(
-                FPTI_ANALYTICS_URL,
-                analyticsRequest.toString(),
-                null,
-                authorization,
-                HttpNoResponse()
+                path = FPTI_ANALYTICS_URL,
+                data = analyticsRequest.toString(),
+                configuration = null,
+                authorization = authorization,
+                callback = HttpNoResponse()
             )
         } catch (e: JSONException) { /* ignored */
         }